change Authentication to a AuthWelcome component screen to include register and sign in buttons
