+++ client/js/screens/AuthWelcome/index.js
@@ -1,0 +1,3 @@
+import AuthWelcomeContainer from "./AuthWelcomeContainer";
+
+export default AuthWelcomeContainer;
