--- client/js/screens/Authentication/Authentication.js
@@ -1,47 +1,0 @@
-import React from "react";
-import { Text, AsyncStorage, View, Button } from "react-native";
-
-class Authentication extends React.Component {
-  static navigationOptions = {
-    title: "SignIn"
-  };
-
-  render() {
-    return (
-      <View>
-        <Button title="Sign in!" onPress={this._signInAsync} />
-      </View>
-    );
-  }
-
-  _signInAsync = async () => {
-    await AsyncStorage.setItem("userToken", "abc");
-    this.props.navigation.navigate("FindLocker");
-  };
-}
-
-class HomeScreen extends React.Component {
-  static navigationOptions = {
-    title: "Welcome to the app!"
-  };
-
-  render() {
-    return (
-      <View>
-        <Button title="Show me more of the app" onPress={this._showMoreApp} />
-        <Button title="Actually, sign me out :)" onPress={this._signOutAsync} />
-      </View>
-    );
-  }
-
-  _showMoreApp = () => {
-    this.props.navigation.navigate("Other");
-  };
-
-  _signOutAsync = async () => {
-    await AsyncStorage.clear();
-    this.props.navigation.navigate("Auth");
-  };
-}
-
-export default Authentication;
