--- client/js/screens/Authentication/index.js
@@ -1,3 +1,0 @@
-import AuthenticationContainer from "./AuthenticationContainer";
-
-export default AuthenticationContainer;
